requirements.txt
# Core dependencies
aiohttp==3.9.1
aiofiles==23.2.1
asyncio
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
click==8.1.7
python-decouple==3.8
certifi

# Scheduling
APScheduler==3.10.4
SQLAlchemy==2.0.23

# Data processing
lxml==4.9.3
html5lib==1.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: For better performance
uvloop==0.19.0  # Unix only